# docs/api/swagger/swagger.yaml
...
swagger: '2.0'
info:
  version: 0.0.1
  title: carmate
  description: Api documentation for carmate
  termsOfService: terms
  contact:
    name: Umer Shehzad
    url: umershehzad.st@gmail.com
    email: umershehzad.st@gmail.com
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
host: 'localhost'
basePath: "/api/v1"
paths:
...
  "/user/":
    get:
      summary: My Profile
      description: This service returns the user profile
      operationId: user-me
      tags:
      - User
      responses:
        '200':
          description: OK
        '401':
          description: Connection Unautherized!
        '404':
          description: Token Not found
        '500':
          description: Internal Server Error
    put:
      summary: Update Profile
      description: This server helps user to update his/her profile
      operationId: user-edit
      tags:
      - User
      parameters:
      - name: body
        in: body
        description: Path the object with username, image, address, phone
        schema:
          type: object
          properties:
            username:
              type: string
              example: umershehzad1
            image:
              type: string
              example: http://localhost:3000/public/upload/default.png
            address:
              type: string
              example: address
            phone:
              type: string
              example: '090078601'
      responses:
        '200':
          description: ok
        '400':
          description: Validation Error
        '401':
          description: Connection Unautherized!
        '404':
          description: Token Not found
        '500':
          description: Internal Server Error
...